+++ Bindings/Python/examples/Moco/examplePredictAndTrack.py
@@ -141,7 +141,7 @@ def solvePrediction():
 
 
     # Configure the solver.
-    solver = study.initTropterSolver()
+    solver = study.initCasADiSolver()
     solver.set_num_mesh_intervals(100)
     solver.set_verbosity(2)
     solver.set_optim_solver("ipopt")
@@ -230,13 +230,10 @@ def solveStateTracking(stateRef):
     # TODO problem.addGoal(effort)
 
     # Configure the solver.
-    solver = study.initTropterSolver()
+    solver = study.initCasADiSolver()
     solver.set_num_mesh_intervals(50)
     solver.set_verbosity(2)
-    solver.set_optim_solver("ipopt")
-    solver.set_optim_jacobian_approximation("exact")
     solver.set_optim_hessian_approximation("exact")
-    solver.set_exact_hessian_block_sparsity_mode("dense")
 
     # Save the problem to a setup file for reference.
     study.printToXML("examplePredictAndTrack_track_states.omoco")
@@ -285,13 +282,10 @@ def solveMarkerTracking(markersRef, guess):
     # problem.addGoal(effort)
 
     # Configure the solver.
-    solver = study.initTropterSolver()
+    solver = study.initCasADiSolver()
     solver.set_num_mesh_intervals(50)
     solver.set_verbosity(2)
-    solver.set_optim_solver("ipopt")
-    solver.set_optim_jacobian_approximation("exact")
     solver.set_optim_hessian_approximation("exact")
-    solver.set_exact_hessian_block_sparsity_mode("dense")
     
     solver.setGuess(guess)